# documentation/openapi.yaml
openapi: 3.0.3
info:
  title: EV Charging Station API
  description: |
    API για τη διαχείριση σταθμών φόρτισης ηλεκτρικών οχημάτων (EV Charging Stations).
    
    ## Αυθεντικοποίηση
    Η πλειοψηφία των endpoints απαιτεί αυθεντικοποίηση μέσω JWT token.
    Το token λαμβάνεται από το endpoint `/api/v1/auth/signin` και πρέπει να 
    συμπεριλαμβάνεται στο header `Authorization` ως `Bearer <token>`.
    
    ## Ρόλοι Χρηστών
    - **USER**: Βασικός χρήστης - μπορεί να δει φορτιστές, να κάνει κρατήσεις, κ.λπ.
    - **ADMIN**: Διαχειριστής - πλήρης πρόσβαση σε όλες τις λειτουργίες διαχείρισης
  version: 1.0.0
  contact:
    name: softeng25-02 Team
servers:
  - url: http://localhost:9876
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token που λαμβάνεται από το signin endpoint
...
paths:
...
  /api/v1/reserve/{id}:
    post:
      tags: [Reservations]
      summary: Κράτηση φορτιστή (30 λεπτά default)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID φορτιστή
          schema:
            type: integer
      responses:
        '200':
          description: Κράτηση επιτυχής
          content:
            application/json:
              schema:
                type: object
                properties:
                  pointid:
                    type: string
                  status:
                    type: string
                    example: "reserved"
                  reservationendtime:
                    type: string
                    description: "Format: YYYY-MM-DD HH:MM"
        '400':
          description: Ο χρήστης έχει ήδη ενεργή κράτηση
        '404':
          description: Φορτιστής δεν βρέθηκε
        '409':
          description: Φορτιστής δεν είναι διαθέσιμος
  /api/v1/reserve/{id}/{minutes}:
    post:
      tags: [Reservations]
      summary: Κράτηση φορτιστή για συγκεκριμένο χρόνο
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: minutes
          in: path
          required: true
          description: Διάρκεια κράτησης (max 60 λεπτά)
          schema:
            type: integer
            maximum: 60
      responses:
        '200':
          description: Κράτηση επιτυχής
          content:
            application/json:
              schema:
                type: object
                properties:
                  pointid:
                    type: string
                  status:
                    type: string
                  reservationendtime:
                    type: string
  /api/v1/reserve/{id}/cancel:
    post:
      tags: [Reservations]
      summary: Ακύρωση κράτησης
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID φορτιστή
          schema:
            type: integer
      responses:
        '200':
          description: Κράτηση ακυρώθηκε
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  chargerId:
                    type: integer
        '404':
          description: Δεν βρέθηκε ενεργή κράτηση

  # ==================== SESSIONS ====================
...